requirements.txt
streamlit
pandas
plotly
pyarrow
fastparquet
pandas-gbq
google-cloud-bigquery
kagglehub
db-dtypes